refactor: rename stream to object_stream
