+++ src/structure/object_stream.rs
